KeyboardChar: store keyChar rather than the event
Creating a "mock" event requires creating a "JComponent", that initializes
the swing ui stuffs, which means that we effectively initialize a swing
"application" when running tests that make use of this class.

This commit is similar to cbd73f4, which solved the same issue for
the MouseButton class and e43060f, which did the same for the KeyboardKey
class.