interaction: implement new mouse handling
Introduce "actionable graphics" that can be acted upon mouse-based
interaction from the user.

Adds three methods to the Graphics class called
withMouse{Press,Release,Drag}Action that take a MouseAction interface
(alias for Function3<T, Coordinate, MouseButton, T>) and a Graphic.

When executing a mouse interaction, first we will try to find an
actionable graphic at the (global) coordinates of the mouse event.
If we find it, we convert the coordinates to coordinates relative to the
actionable graphic itself, and execute the "action".
If there's no actionable graphic at that location, we fall back to the
existing "global" coordinate handlers.

Note that since d2a16d0, the background is rendered separatedly on a
different canvas, so we won't look for actionables in the background
of an interaction.