interaction: render background separatedly
This allows us to save time not re-rendering the background at every
tick / input event.
As a consequence of this change, the background and the foreground are
no longer "composed", but they are instead drawn by two different
canvases which are aligned to the top-left (coordinate 0,0 of their
parent JPanel). This will require the foreground and background to be
"manually" aligned by the user by making them of the same size.